vendor_management_system/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-model-utils==4.5.1
djangorestframework==3.15.1
Faker==25.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
